Added a 128x96 Textmode renderer using 8x8 tiles on an 1024x768 SVGA screen. You can change resolution by modifying boot.S and textDISP.hpp

+++ kernel/hal/x86/include/base.hpp
@@ -13,8 +13,8 @@ namespace MTGosHAL {
 	class BlockDevice;
 	class Task;
 	class VMM3;
-	enum class BG_color: uint16_t;
-	enum class FG_color: uint16_t;
+	enum class BG_color: uint32_t;
+	enum class FG_color: uint32_t;
 	extern Serial debug;
 	extern Screen out;
 	extern Screen err;
